pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adaptive-agent-mcp"
version = "0.3.0"
description = "让你的 AI 真正记住你 — 基于 MCP 协议的跨应用持久化记忆系统"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "justforever17" }
]
keywords = [
    "mcp",
    "memory",
    "ai-agent",
    "llm",
    "persistent-memory",
    "cross-application"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "mcp>=1.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    "filelock>=3.12.0"
]

[project.urls]
Homepage = "https://github.com/justforever17/adaptive-agent-mcp"
Repository = "https://github.com/justforever17/adaptive-agent-mcp"
Documentation = "https://github.com/justforever17/adaptive-agent-mcp#readme"

[project.scripts]
adaptive-agent-mcp = "adaptive_agent_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["adaptive_agent_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "adaptive_agent_mcp/**/*.py",
    "README.md",
    "LICENSE"
]
exclude = [
    "**/__pycache__",
    "**/*.pyc",
    "docs/",
    ".agent/",
    ".venv/"
]
